[FIX] NavigationItem 의존성 제거
